# openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "EvalHarness API",
  "description": "Runs agentic tool-use scenarios against any LiteLLM-supported model and grades them deterministically.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/runs": {
   "post": {
    "tags": [
     "runs"
    ],
    "summary": "Create a run and queue it for execution",
    "operationId": "createRun",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RunCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "runs"
    ],
    "summary": "Paginated runs, filterable by model and status",
    "operationId": "listRuns",
    "parameters": [
     {
      "name": "model_key",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Model Key"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/RunStatus"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 25,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunListResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/runs/{run_id}": {
   "get": {
    "tags": [
     "runs"
    ],
    "summary": "One run with its attempts and summary",
    "operationId": "getRun",
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/runs/{run_id}/cancel": {
   "post": {
    "tags": [
     "runs"
    ],
    "summary": "Cancel a queued or running run",
    "description": "Set the cancellation flag that attempts check between model calls.\n\nThe flag lives in two places on purpose: Redis, which a running attempt polls\ncheaply, and the run's status in Postgres, which is authoritative and\nsurvives a Redis restart.",
    "operationId": "cancelRun",
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CancelResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/runs/{run_id}/stream": {
   "get": {
    "tags": [
     "runs"
    ],
    "summary": "Server-Sent Events: attempt started/finished, run completed",
    "operationId": "streamRun",
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "A text/event-stream whose data frames are ProgressEvent JSON.",
      "content": {
       "text/event-stream": {},
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProgressEvent"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AttemptStatus": {
    "type": "string",
    "enum": [
     "queued",
     "running",
     "completed",
     "failed",
     "cancelled"
    ],
    "title": "AttemptStatus"
   },
   "AttemptSummary": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "run_id": {
      "type": "string",
      "title": "Run Id"
     },
     "scenario_id": {
      "type": "string",
      "title": "Scenario Id"
     },
     "repetition": {
      "type": "integer",
      "title": "Repetition"
     },
     "status": {
      "$ref": "#/components/schemas/AttemptStatus"
     },
     "passed": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Passed"
     },
     "critical_failure": {
      "type": "boolean",
      "title": "Critical Failure",
      "default": false
     },
     "axis_scores": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Axis Scores"
     },
     "cost_usd": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cost Usd"
     },
     "input_tokens": {
      "type": "integer",
      "title": "Input Tokens",
      "default": 0
     },
     "output_tokens": {
      "type": "integer",
      "title": "Output Tokens",
      "default": 0
     },
     "duration_ms": {
      "type": "integer",
      "title": "Duration Ms",
      "default": 0
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "id",
     "run_id",
     "scenario_id",
     "repetition",
     "status"
    ],
    "title": "AttemptSummary"
   },
...
   "CancelResponse": {
    "properties": {
     "run": {
      "$ref": "#/components/schemas/RunDetail"
     },
     "cancelled_attempts": {
      "type": "integer",
      "title": "Cancelled Attempts",
      "default": 0
     },
     "already_terminal": {
      "type": "boolean",
      "title": "Already Terminal",
      "default": false
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "run"
    ],
    "title": "CancelResponse"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "JsonValue": {},
...
   "ProgressEvent": {
    "properties": {
     "type": {
      "type": "string",
      "enum": [
       "snapshot",
       "run_started",
       "attempt_started",
       "attempt_finished",
       "run_completed",
       "heartbeat"
      ],
      "title": "Type"
     },
     "run_id": {
      "type": "string",
      "title": "Run Id"
     },
     "run_status": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/RunStatus"
       },
       {
        "type": "null"
       }
      ]
     },
     "attempt_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Attempt Id"
     },
     "scenario_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Scenario Id"
     },
     "repetition": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Repetition"
     },
     "attempt_status": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/AttemptStatus"
       },
       {
        "type": "null"
       }
      ]
     },
     "passed": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Passed"
     },
     "completed_attempts": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Completed Attempts"
     },
     "total_attempts": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Total Attempts"
     },
     "attempt_counts": {
      "additionalProperties": {
       "type": "integer"
      },
      "type": "object",
      "title": "Attempt Counts"
     },
     "message": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Message"
     },
     "ts": {
      "type": "string",
      "format": "date-time",
      "title": "Ts"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "type",
     "run_id",
     "ts"
    ],
    "title": "ProgressEvent",
    "description": "One Server-Sent Event on ``GET /api/runs/{id}/stream`` (SPEC 9.1).\n\nEmitted as ``event: <type>`` with this object as the JSON ``data`` payload."
   },
...
   "Run": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "status": {
      "$ref": "#/components/schemas/RunStatus"
     },
     "model_key": {
      "type": "string",
      "title": "Model Key"
     },
     "litellm_model": {
      "type": "string",
      "title": "Litellm Model"
     },
     "params": {
      "additionalProperties": {
       "$ref": "#/components/schemas/JsonValue"
      },
      "type": "object",
      "title": "Params"
     },
     "k": {
      "type": "integer",
      "title": "K"
     },
     "scenario_ids": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Scenario Ids"
     },
     "config_hashes": {
      "additionalProperties": {
       "type": "string"
      },
      "type": "object",
      "title": "Config Hashes"
     },
     "git_commit": {
      "type": "string",
      "title": "Git Commit",
      "default": "unknown"
     },
     "harness_version": {
      "type": "string",
      "title": "Harness Version",
      "default": "0.0.0"
     },
     "summary": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/RunSummary"
       },
       {
        "type": "null"
       }
      ]
     },
     "attempt_counts": {
      "additionalProperties": {
       "type": "integer"
      },
      "type": "object",
      "title": "Attempt Counts",
      "description": "Attempt count by status."
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "id",
     "created_at",
     "status",
     "model_key",
     "litellm_model",
     "k"
    ],
    "title": "Run",
    "description": "A run without its (large) scenario snapshots."
   },
   "RunCreate": {
    "properties": {
     "model_key": {
      "type": "string",
      "title": "Model Key"
     },
     "scenario_ids": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "minItems": 1,
      "title": "Scenario Ids"
     },
     "k": {
      "type": "integer",
      "maximum": 50.0,
      "minimum": 1.0,
      "title": "K",
      "default": 1
     },
     "params_override": {
      "anyOf": [
       {
        "additionalProperties": {
         "$ref": "#/components/schemas/JsonValue"
        },
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Params Override"
     },
     "transcript": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Transcript",
      "description": "FakeModel only: transcript stem to replay (default 'golden')."
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "model_key",
     "scenario_ids"
    ],
    "title": "RunCreate",
    "description": "Body of ``POST /api/runs`` (SPEC 9.2)."
   },
   "RunDetail": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "status": {
      "$ref": "#/components/schemas/RunStatus"
     },
     "model_key": {
      "type": "string",
      "title": "Model Key"
     },
     "litellm_model": {
      "type": "string",
      "title": "Litellm Model"
     },
     "params": {
      "additionalProperties": {
       "$ref": "#/components/schemas/JsonValue"
      },
      "type": "object",
      "title": "Params"
     },
     "k": {
      "type": "integer",
      "title": "K"
     },
     "scenario_ids": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Scenario Ids"
     },
     "config_hashes": {
      "additionalProperties": {
       "type": "string"
      },
      "type": "object",
      "title": "Config Hashes"
     },
     "git_commit": {
      "type": "string",
      "title": "Git Commit",
      "default": "unknown"
     },
     "harness_version": {
      "type": "string",
      "title": "Harness Version",
      "default": "0.0.0"
     },
     "summary": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/RunSummary"
       },
       {
        "type": "null"
       }
      ]
     },
     "attempt_counts": {
      "additionalProperties": {
       "type": "integer"
      },
      "type": "object",
      "title": "Attempt Counts",
      "description": "Attempt count by status."
     },
     "attempts": {
      "items": {
       "$ref": "#/components/schemas/AttemptSummary"
      },
      "type": "array",
      "title": "Attempts"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "id",
     "created_at",
     "status",
     "model_key",
     "litellm_model",
     "k"
    ],
    "title": "RunDetail"
   },
   "RunListResponse": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/Run"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "limit": {
      "type": "integer",
      "title": "Limit"
     },
     "offset": {
      "type": "integer",
      "title": "Offset"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "items",
     "total",
     "limit",
     "offset"
    ],
    "title": "RunListResponse"
   },
   "RunStatus": {
    "type": "string",
    "enum": [
     "queued",
     "running",
     "completed",
     "failed",
     "cancelled"
    ],
    "title": "RunStatus"
   },
   "RunSummary": {
    "properties": {
     "k": {
      "type": "integer",
      "title": "K",
      "default": 1
     },
     "total_attempts": {
      "type": "integer",
      "title": "Total Attempts",
      "default": 0
     },
     "completed_attempts": {
      "type": "integer",
      "title": "Completed Attempts",
      "default": 0
     },
     "passed_attempts": {
      "type": "integer",
      "title": "Passed Attempts",
      "default": 0
     },
     "failed_attempts": {
      "type": "integer",
      "title": "Failed Attempts",
      "default": 0
     },
     "cancelled_attempts": {
      "type": "integer",
      "title": "Cancelled Attempts",
      "default": 0
     },
     "errored_attempts": {
      "type": "integer",
      "title": "Errored Attempts",
      "default": 0
     },
     "pass_at_1": {
      "type": "number",
      "title": "Pass At 1",
      "default": 0.0
     },
     "pass_hat_k": {
      "type": "number",
      "title": "Pass Hat K",
      "default": 0.0
     },
     "axis_scores": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Axis Scores"
     },
     "cost_usd": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cost Usd",
      "description": "Null when any attempt's cost is unknown; never guessed."
     },
     "latency_p50_ms": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Latency P50 Ms"
     },
     "latency_p95_ms": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Latency P95 Ms"
     },
     "model_calls": {
      "type": "integer",
      "title": "Model Calls",
      "default": 0
     },
     "input_tokens": {
      "type": "integer",
      "title": "Input Tokens",
      "default": 0
     },
     "output_tokens": {
      "type": "integer",
      "title": "Output Tokens",
      "default": 0
     },
     "mean_steps_per_turn": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mean Steps Per Turn"
     },
     "per_scenario": {
      "items": {
       "$ref": "#/components/schemas/ScenarioStats"
      },
      "type": "array",
      "title": "Per Scenario"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "RunSummary",
    "description": "The ``runs.summary`` jsonb blob, computed by ``finalize_run`` (SPEC 6.3)."
   },
...
   "ScenarioStats": {
    "properties": {
     "scenario_id": {
      "type": "string",
      "title": "Scenario Id"
     },
     "config_hash": {
      "type": "string",
      "title": "Config Hash",
      "default": ""
     },
     "attempts": {
      "type": "integer",
      "title": "Attempts",
      "default": 0
     },
     "completed": {
      "type": "integer",
      "title": "Completed",
      "default": 0
     },
     "passed": {
      "type": "integer",
      "title": "Passed",
      "default": 0
     },
     "pass_at_1": {
      "type": "number",
      "title": "Pass At 1",
      "default": 0.0
     },
     "pass_hat_k": {
      "type": "number",
      "title": "Pass Hat K",
      "default": 0.0
     },
     "axis_scores": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Axis Scores"
     },
     "cost_usd": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cost Usd"
     },
     "latency_p50_ms": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Latency P50 Ms"
     },
     "latency_p95_ms": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Latency P95 Ms"
     },
     "input_tokens": {
      "type": "integer",
      "title": "Input Tokens",
      "default": 0
     },
     "output_tokens": {
      "type": "integer",
      "title": "Output Tokens",
      "default": 0
     },
     "mean_steps_per_turn": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mean Steps Per Turn"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "scenario_id"
    ],
    "title": "ScenarioStats",
    "description": "Per-scenario slice of a run summary (SPEC 6.3)."
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}