# web/lib/vocabulary.ts
/**
 * The parts of the domain vocabulary that must exist at runtime.
 *
 * Response *shapes* come from `lib/api-types.ts` (generated by
 * `npm run gen:api`) and are re-exported through `lib/types.ts`. But the UI
 * also needs to *iterate* the axes and statuses — to lay out a radar chart, or
 * to build a status filter — and a generated union is erased at runtime. So
 * these lists live here, mirroring `backend/evalharness/schema/enums.py`.
 *
 * `lib/types.ts` checks each list against the generated union, so if the
 * backend adds an axis and this file does not, the build fails.
 */

export type JsonValue = string | number | boolean | null | JsonValue[] | { [key: string]: JsonValue };

export type JsonObject = { [key: string]: JsonValue };

export const AXES = [
  "selection",
  "arguments",
  "ordering",
  "restraint",
  "recovery",
  "state",
  "clarification",
  "safety",
] as const;
export type Axis = (typeof AXES)[number];

export const SEVERITIES = ["critical", "required", "soft"] as const;
export type Severity = (typeof SEVERITIES)[number];

export const SCOPES = ["turn", "scenario"] as const;
export type Scope = (typeof SCOPES)[number];

export const ASSERTION_TYPES = [
  "tool_called",
  "tool_not_called",
  "order",
  "parallel",
  "no_tool_calls",
  "clarification",
  "recovered",
  "tool_result_matches",
  "args_not_contains",
  "response_matches",
  "response_not_matches",
  "judge",
] as const;
export type AssertionType = (typeof ASSERTION_TYPES)[number];

export const EVENT_TYPES = [
  "system",
  "user_message",
  "model_request",
  "model_response",
  "tool_call",
  "tool_result",
  "fault",
  "retry",
  "limit_exceeded",
  "error",
] as const;
export type EventType = (typeof EVENT_TYPES)[number];

export const RUN_STATUSES = ["queued", "running", "completed", "failed", "cancelled"] as const;
export type RunStatus = (typeof RUN_STATUSES)[number];

export const ATTEMPT_STATUSES = ["queued", "running", "completed", "failed", "cancelled"] as const;
export type AttemptStatus = (typeof ATTEMPT_STATUSES)[number];

export type MockKind = "fixture" | "handler";